+++ src/presets.ts
@@ -2,5 +2,4 @@ import { ContainerSettings } from './types';
 
 export const DEFAULT_SETTINGS: Required<ContainerSettings> = {
     allowRewriteDependencies: false,
-    resolveInContainerContext: false,
 };